finally make Layouts work
this is so convulated and gay it really needs work.

but so does the networking and basically everything else so... as long as it works for now.

I think after this I can get to making things focused and selectable, and then InputBox will finally only accept input while it's focused. then user accounts and proper functions can be added.